+++ src/valueObjects/Offset.java
@@ -1,6 +1,5 @@
 package valueObjects;
 
-import java.math.BigDecimal;
 
 /**
  * Offsets a value object by a certain amount of ticks.
@@ -21,7 +20,7 @@ public class Offset implements IValued {
 	}
 	
 	@Override
-	public BigDecimal getValue(int ticksBack) throws TickOutOfRangeException {
+	public double getValue(int ticksBack) throws TickOutOfRangeException {
 		//Return the value offset ticks ago.
 		return underlying.getValue(ticksBack + offset);
 	}
